requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
genai-prices==0.0.52
griffe==1.15.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.12.0
logfire-api==4.21.0
MarkupSafe==3.0.3
openai==2.16.0
opentelemetry-api==1.39.1
pydantic==2.12.5
pydantic-ai-slim==1.48.0
pydantic-graph==1.48.0
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
zipp==3.23.0
